fix integer overflow for FSK bitrate
+ add support for sx127x power profiling
+ better logging
+ fix restart CAD mode

+++ components/sx127x_util/sx127x_util.h
@@ -45,8 +45,8 @@ typedef struct {
   uint64_t endTimeMillis;
   uint64_t currentTimeMillis;
   uint64_t freq;
-  uint16_t bitrate;
-  uint16_t freq_deviation;
+  uint32_t bitrate;
+  uint32_t freq_deviation;
   uint16_t preamble;
   uint8_t *syncword;
   uint8_t syncword_length;
